Rename models to shared-models
