+++ bot/classes/owner.js
@@ -1,4 +1,4 @@
-const OwnerModel = require("./../../models/owner.js");
+const OwnerModel = require("./../../shared-models/owner.js");
 const webAPI = require("../../dashboard/backend/api.js").getWebAPI("discotron-dashboard");
 const db = require("./../apis/database-crud.js");
 const Logger = require("../utils/logger.js");
