+++ bot/classes/user-role.js
@@ -1,4 +1,4 @@
-const UserRoleModel = require("./../../models/user-role.js");
+const UserRoleModel = require("./../../shared-models/user-role.js");
 const db = require("./../apis/database-crud.js");
 const Logger = require("../utils/logger.js");
 const discordClientProvider = require("./../apis/discord-client-provider.js");
