+++ dashboard/www/index.html
@@ -64,14 +64,14 @@
     <script src="/dashboard/scripts/mobile-helper.js"></script>
 
     <!-- Models definition -->
-    <script src="/models/bot-settings.js"></script>
-    <script src="/models/command.js"></script>
-    <script src="/models/guild.js"></script>
-    <script src="/models/owner.js"></script>
-    <script src="/models/permission.js"></script>
-    <script src="/models/plugin.js"></script>
-    <script src="/models/repository.js"></script>
-    <script src="/models/user-role.js"></script>
+    <script src="/shared-models/bot-settings.js"></script>
+    <script src="/shared-models/command.js"></script>
+    <script src="/shared-models/guild.js"></script>
+    <script src="/shared-models/owner.js"></script>
+    <script src="/shared-models/permission.js"></script>
+    <script src="/shared-models/plugin.js"></script>
+    <script src="/shared-models/repository.js"></script>
+    <script src="/shared-models/user-role.js"></script>
 
     <!-- Models -->
     <script src="/dashboard/scripts/models/channel.js"></script>
