+++ shared-models/bot-settings.js
