+++ shared-models/command.js
